conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asn1crypto=0.22.0=py36h265ca7c_1
bzip2=1.0.6=h0376d23_1
ca-certificates=2017.08.26=h1d4fec5_0
certifi=2017.7.27.1=py36h8b7b77e_0
cffi=1.10.0=py36had8d393_1
chardet=3.0.4=py36h0f667ec_1
cryptography=2.0.3=py36ha225213_1
cudatoolkit=8.0=3
cudnn=6.0.21=cuda8.0_0
cycler=0.10.0=py36h93f1223_0
dbus=1.10.22=h3b5a359_0
expat=2.2.4=h6ea4f2b_2
fontconfig=2.12.4=h88586e7_1
freetype=2.8=h52ed37b_0
fuel=0.2.0=py36_0
glib=2.53.6=h5d9569c_2
gst-plugins-base=1.12.2=he3457e5_0
gstreamer=1.12.2=h4f93127_0
h5py=2.7.0=py36he81ebca_1
hdf5=1.10.1=hb0523eb_0
icu=58.2=h211956c_0
idna=2.6=py36h82fb2a8_1
intel-openmp=2018.0.0=h15fc484_7
jpeg=9b=habf39ab_1
libedit=3.1=heed3624_0
libffi=3.2.1=h4deb6c0_3
libgcc=7.2.0=h69d50b8_2
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran=3.0.0=1
libgfortran-ng=7.2.0=h9f7466a_2
libpng=1.6.32=hda9c8bc_2
libsodium=1.0.13=h31c71d8_2
libstdcxx-ng=7.2.0=h7a57d05_2
libtiff=4.0.8=h90200ff_9
libxcb=1.12=h84ff03f_3
libxml2=2.9.4=h6b072ca_5
lzo=2.10=h1bfc0ba_1
matplotlib=2.1.0=py36hba5de38_0
mkl=2018.0.1=h19d6760_4
nccl=1.3.4=cuda8.0_1
ncurses=6.0=h06874d7_1
numexpr=2.6.4=py36hc4a3f9a_0
numpy=1.13.3=py36ha12f23b_0
olefile=0.44=py36h79f9f78_0
openssl=1.0.2l=h9d1a558_3
pcre=8.41=hc71a17e_0
picklable-itertools=0.1.1=py36_0
pillow=4.2.1=py36h9119f52_0
pip=9.0.1=py36h8ec8b28_3
progressbar2=3.18.1=py36_0
py=1.4.34=py36h0712aa3_1
pycparser=2.18=py36hf9f622e_1
pyopenssl=17.2.0=py36h5cc804b_0
pyparsing=2.2.0=py36hee85983_1
pyqt=5.6.0=py36h0386399_5
pysocks=1.6.7=py36hd97a5b1_1
pytables=3.4.2=py36h3b5282a_2
pytest=3.2.1=py36h11ad3bb_1
pytest-runner=2.11.1=py36h66b2b85_0
python=3.6.3=hcad60d5_0
python-dateutil=2.6.1=py36h88d3b88_1
python-utils=2.2.0=py36_0
pytorch=0.3.0=py36_cuda8.0.61_cudnn7.0.3h37a80b5_4
pytz=2017.2=py36hc2ccc2a_1
pyyaml=3.12=py36hafb9ca4_1
pyzmq=16.0.2=py36h3b0cf96_2
qt=5.6.2=h974d657_12
readline=7.0=hac23ff0_3
requests=2.18.4=py36he2e5f8d_1
scipy=1.0.0=py36hbf646e7_0
setuptools=36.5.0=py36he42e2e1_0
sip=4.18.1=py36h51ed4ed_2
six=1.10.0=py36hcac75e4_1
sqlite=3.20.1=h6d8b0f3_1
tk=8.6.7=h5979e9b_1
torchvision=0.2.0=py36h17b6947_1
tornado=4.5.2=py36h1283b2a_0
tqdm=4.15.0=py36_0
urllib3=1.22=py36hbe7ace6_0
wheel=0.29.0=py36he7f4e38_1
xz=5.2.3=h2bcbf08_1
yaml=0.1.7=h96e3832_1
zeromq=4.2.2=hb0b69da_1
zlib=1.2.11=hfbfcf68_1
